Updated API, added two commands and State+Event
git-svn-id: file:///home/srv/www/butt.se/pnp/svn/trunk@4 f672c3ff-8b41-4f22-a4ab-2adcb4f732c7

+++ MachineAPI/MachineAPI/RotateAbsolute.cs
@@ -4,9 +4,16 @@ using System.Text;
 
 namespace MachineController
 {
+	/// <summary>
+	/// Rotate to an absolute angle
+	/// </summary>
 	public class RotateAbsolute : Command
 	{
-		public RotateAbsolute(int angle)
+		/// <summary>
+		/// Rotate 
+		/// </summary>
+		/// <param name="angle">The angle in degrees</param>
+		public RotateAbsolute(float angle)
 		{
 			throw new System.NotImplementedException();
 		}
